Implemented a basic design for gsoc project
