+++ examples/global_init/CMakeLists.txt
@@ -7,8 +7,8 @@
 
 add_hpx_executable(global_factorial
   SOURCES factorial.cpp
-  FOLDER "Examples/Glob_init/")
+  FOLDER "Examples/Global_init/")
 
-add_hpx_pseudo_target(examples.glob_init.global_factorial)
-add_hpx_pseudo_dependencies(examples.glob_init examples.glob_init.global_factorial)
-add_hpx_pseudo_dependencies(examples.glob_init.global_factorial global_factorial_exe)
+add_hpx_pseudo_target(examples.global_init.global_factorial)
+add_hpx_pseudo_dependencies(examples.global_init examples.global_init.global_factorial)
+add_hpx_pseudo_dependencies(examples.global_init.global_factorial global_factorial_exe)
